Split game into Overworld an Battle
